requirements.txt
alabaster==0.7.13
appdirs==1.4.4
Babel==2.13.1
bootstrap-py==2.0.0
certifi==2023.7.22
charset-normalizer==3.3.1
colorama==0.4.6
defusedxml==0.7.1
docutils==0.20.1
gitdb==4.0.11
GitPython==3.1.40
idna==3.4
imagesize==1.4.1
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
packaging==23.2
pguard==0.4.0
pipdate==0.5.6
Pygments==2.16.1
requests==2.31.0
rich==13.6.0
setuptools-scm==8.0.4
smmap==5.0.1
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
typing_extensions==4.8.0
urllib3==2.0.7
Flask==2.2.5
Flask-Restful==0.3.9
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
